chatbot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
flask==2.3.3
python-docx==1.0.1
PyPDF2==3.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
gunicorn==21.2.0
pymongo==4.6.1
flask-cors==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0 